Refactor: Relocate security and crypto modules
Moved all security and cryptography-related files from `sdk/src/commonMain/kotlin/com/zeropay/sdk/crypto` and `sdk/src/commonMain/kotlin/com/zeropay/sdk/security` to a new `sdk/src/androidMain/kotlin/com/zeropay/sdk/security` directory. This consolidates platform-specific security implementations.

Additionally, relocated all UI tests from `sdk/src/test` to `sdk/src/androidTest` to properly categorize them as instrumented tests.

Updated all import statements across the `sdk`, `enrollment`, and `merchant` modules to reflect these changes. Also included minor code cleanup and bug fixes in related files.

+++ sdk/src/androidMain/kotlin/com/zeropay/sdk/ui/ImageTapCanvas.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.zeropay.sdk.crypto.CryptoUtils
+import com.zeropay.sdk.security.CryptoUtils
 import com.zeropay.sdk.factors.ImageTapFactor
 
 /**
@@ -140,7 +140,7 @@ fun ImageTapCanvas(
                             
                             // Store image size on first tap
                             if (imageSize == null) {
-                                imageSize = size.width to size.height
+                                imageSize = size.width.toFloat() to size.height.toFloat()
                             }
                             
                             // Validate tap distance from previous taps
